bring mocked Mockaroo data out into its own file
